+++ src/main/java/ru/tinkoff/rentall/repository/AdvertisementReviewRepository.java
@@ -3,6 +3,6 @@ package ru.tinkoff.rentall.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.tinkoff.rentall.entity.AdvertisementReview;
 
-public interface AdvertisementReviewRepository extends JpaRepository<AdvertisementReview, Long> {
+public interface AdvertisementReviewRepository extends JpaRepository<AdvertisementReview, Integer> {
 
 }
