+++ src/main/java/ru/tinkoff/rentall/repository/UserReviewRepository.java
@@ -3,6 +3,6 @@ package ru.tinkoff.rentall.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.tinkoff.rentall.entity.UserReview;
 
-public interface UserReviewRepository extends JpaRepository<UserReview, Long> {
+public interface UserReviewRepository extends JpaRepository<UserReview, Integer> {
 
 }
